refactor header actions into single components
