+++ components/Layout/header/LayoutHeaderAccount.vue
@@ -1,0 +1,44 @@
+<script setup lang="ts">
+const customerStore = useCustomerStore();
+const { signedIn } = storeToRefs(customerStore);
+const modalController = useModal();
+</script>
+
+<template>
+    <LazySharedModal
+        v-if="!signedIn"
+        :with-close-button="true"
+    >
+        <template #trigger>
+            <FormKitIcon
+                class="block h-6 w-6"
+                icon="user"
+            />
+        </template>
+        <template #title>Login</template>
+        <template #content>
+            <AccountLoginRegisterTabs />
+        </template>
+    </LazySharedModal>
+    <LazySharedPopover v-else>
+        <template #trigger>
+            <FormKitIcon
+                class="block h-6 w-6"
+                icon="user"
+                @click="!signedIn ? modalController.open() : null"
+            />
+        </template>
+        <template #content>
+            <div class="py-2 first:pt-0">
+                <NuxtLink to="/account">account</NuxtLink>
+            </div>
+            <FormKit
+                type="submit"
+                prefix-icon="right-from-bracket"
+                @click.prevent="customerStore.logout()"
+            >
+                logout
+            </FormKit>
+        </template>
+    </LazySharedPopover>
+</template>
